restructure files and folders
