+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Navbar from './components/Navbar';
-import Home from './components/pages/home/Home';
-import About from './components/pages/about/About';
-import Contact from './components/pages/contact/Contact';
+import Navbar from './components/bars/Navbar';
+import Home from './pages/home/Home';
+import About from './pages/about/About';
+import Contact from './pages/contact/Contact';
 import './App.css';
 
 function App() {
